+++ tasks/mpi/voroshilov_v_torus_grid/perf_tests/main.cpp
@@ -125,13 +125,13 @@ TEST(voroshilov_v_torus_grid_mpi_perf, test_pipeline_run_mpi) {
   // ASSERT_EQ() doesnt work if world.rank() != 0 so decided send to process 0 and check there
   if (world.rank() == dst_proc) {
     bool flag_data = true;
-    for (int i = 0; i < output_data.size(); i++) {
+    for (size_t i = 0; i < output_data.size(); i++) {
       if (output_data[i] != input_data[i]) {
         flag_data = false;
       }
     }
     bool flag_path = true;
-    for (int i = 0; i < output_path.size(); i++) {
+    for (size_t i = 0; i < output_path.size(); i++) {
       if (output_path[i] != expected_path[i]) {
         flag_path = false;
       }
@@ -227,13 +227,13 @@ TEST(voroshilov_v_torus_grid_mpi_perf, test_task_run_mpi) {
   // ASSERT_EQ() doesnt work if world.rank() != 0 so decided send to process 0 and check there
   if (world.rank() == dst_proc) {
     bool flag_data = true;
-    for (int i = 0; i < output_data.size(); i++) {
+    for (size_t i = 0; i < output_data.size(); i++) {
       if (output_data[i] != input_data[i]) {
         flag_data = false;
       }
     }
     bool flag_path = true;
-    for (int i = 0; i < output_path.size(); i++) {
+    for (size_t i = 0; i < output_path.size(); i++) {
       if (output_path[i] != expected_path[i]) {
         flag_path = false;
       }
